Revert "move middleware (#22)"
This reverts commit e51e4627ceb7230b09f729590f26bf409725a6f3.
