+++ authentication/main.go
@@ -1,14 +1,13 @@
 package main
 
 import (
-  "fmt"
-  "time"
-  "github.com/dgrijalva/jwt-go"
-  "net/http"
-  "github.com/gin-contrib/cors"
-  "github.com/gin-gonic/gin"
-  "day-trader/middleware"
- )
+	"fmt"
+	"github.com/dgrijalva/jwt-go"
+	"github.com/gin-contrib/cors"
+	"github.com/gin-gonic/gin"
+	"net/http"
+	"time"
+)
 
 // TODO: need env to store secret key
 var secretKey = []byte("secret")
@@ -148,6 +147,6 @@ func main() {
 
 	router.POST("/login", postLogin)
 	router.POST("/register", postRegister)
-	router.GET("/eatCookies", identification.AuthMiddlewareTest, getCookies)
+	router.GET("/eatCookies", getCookies)
 	router.Run(":8888")
 }
