--- middleware/go.mod
@@ -1,3 +1,0 @@
-module day-trader/middleware
-
-go 1.21.6
