--- test/go.mod
@@ -1,3 +1,0 @@
-module day-trader/test
-
-go 1.21.6
