--- test/test.go
@@ -1,9 +1,0 @@
-package test
-
-import (
-	"fmt"
-	)
-
-func Test() {
-	fmt.Println("Test Middleware: ")
-}
