+++ transaction/main.go
@@ -35,9 +35,9 @@ func handleError(c *gin.Context, statusCode int, message string, err error) {
 }
 
 func addMoneyToWallet(c *gin.Context) {
-	user_name, err := c.Get("user_name")
-	
-	if !err {
+	userName, _ := c.Get("user_name")
+
+	if userName == nil {
 		handleError(c, http.StatusBadRequest, "Failed to obtain the user name", nil)
 		return
 	}
