+++ transaction/middleware/identification.go
@@ -1,4 +1,4 @@
-package identification
+package middleware
 
 import (
 	"fmt"
@@ -74,14 +74,5 @@ func Identification(c *gin.Context) {
 	fmt.Println(claims.ExpiresAt)
 	fmt.Println(claims.UserName)
 	c.Set("user_name", claims.UserName)
-	c.Next()
-}
-
-func AuthTest() {
-	fmt.Println("Test Auth: ")
-}
-
-func AuthMiddlewareTest(c *gin.Context) {
-	fmt.Println("Test Auth Middleware: ")
 	c.Next()
 }
