+++ transaction/middleware/test.go
@@ -1,0 +1,11 @@
+package middleware
+
+import (
+	"fmt"
+	"github.com/gin-gonic/gin"
+	)
+
+func Test(c *gin.Context) {
+	fmt.Println("Test Middleware: ")
+	c.Next()
+}
